e2e: remove activator container and consolidate to onchain allocation
Drops the legacy-allocation BackwardCompatibility shard, floors the
surviving (onchain) BackwardCompatibility shard at CLI 0.12.0, and
removes the activator container from the e2e devnet entirely. All
remaining shards now run with onchain allocation; the
DZ_E2E_ONCHAIN_ALLOCATION and DZ_E2E_DISABLE_ACTIVATOR env knobs are
gone.

Closes #3609
Closes #3610
Closes #3611
