+++ .github/workflows/e2e.yml
@@ -17,7 +17,6 @@ env:
   DZ_BASE_IMAGE: ghcr.io/malbeclabs/dz-e2e/base:${{ github.sha }}
   DZ_LEDGER_IMAGE: ghcr.io/malbeclabs/dz-e2e/ledger:${{ github.sha }}
   DZ_CONTROLLER_IMAGE: ghcr.io/malbeclabs/dz-e2e/controller:${{ github.sha }}
-  DZ_ACTIVATOR_IMAGE: ghcr.io/malbeclabs/dz-e2e/activator:${{ github.sha }}
   DZ_MANAGER_IMAGE: ghcr.io/malbeclabs/dz-e2e/manager:${{ github.sha }}
   DZ_FUNDER_IMAGE: ghcr.io/malbeclabs/dz-e2e/funder:${{ github.sha }}
   DZ_DEVICE_IMAGE: ghcr.io/malbeclabs/dz-e2e/device:${{ github.sha }}
@@ -57,7 +56,6 @@ jobs:
           docker push ${{ env.DZ_IMAGE_REPO }}/base:${{ env.DZ_IMAGE_TAG }}
           docker push ${{ env.DZ_IMAGE_REPO }}/ledger:${{ env.DZ_IMAGE_TAG }}
           docker push ${{ env.DZ_IMAGE_REPO }}/controller:${{ env.DZ_IMAGE_TAG }}
-          docker push ${{ env.DZ_IMAGE_REPO }}/activator:${{ env.DZ_IMAGE_TAG }}
           docker push ${{ env.DZ_IMAGE_REPO }}/manager:${{ env.DZ_IMAGE_TAG }}
           docker push ${{ env.DZ_IMAGE_REPO }}/funder:${{ env.DZ_IMAGE_TAG }}
           docker push ${{ env.DZ_IMAGE_REPO }}/device:${{ env.DZ_IMAGE_TAG }}
@@ -84,9 +82,6 @@ jobs:
           # many parallel subtests internally and would starve other tests.
           dedicated="TestE2E_BackwardCompatibility"
 
-          # Shard 2: Tests that also run with onchain allocation enabled.
-          onchain_tests="TestE2E_IBRL|TestE2E_IBRL_WithAllocatedIP|TestE2E_Multicast|TestE2E_InterfaceValidation|TestE2E_UserLimits|TestE2E_MulticastPublisher_BothAllocationPaths"
-
           remaining=$(echo "$tests" | grep -v "^${dedicated}$")
 
           # Distribute remaining tests round-robin across shards
@@ -106,13 +101,10 @@ jobs:
             i=$((i + 1))
           done <<< "$remaining"
 
-          # Build JSON matrix: shard 1 = dedicated (legacy), shard 2 = dedicated (onchain),
-          # shard 3 = onchain (no activator), shards 4+ = round-robin (legacy)
-          matrix="[{\"shard\":1,\"run\":\"^(${dedicated})$\",\"onchain_allocation\":\"\",\"disable_activator\":\"\"}"
-          matrix="${matrix},{\"shard\":2,\"run\":\"^(${dedicated})$\",\"onchain_allocation\":\"true\",\"disable_activator\":\"\"}"
-          matrix="${matrix},{\"shard\":3,\"run\":\"^(${onchain_tests})$\",\"onchain_allocation\":\"true\",\"disable_activator\":\"true\"}"
+          # Build JSON matrix: shard 1 = BackwardCompatibility, shards 2+ = round-robin
+          matrix="[{\"shard\":1,\"run\":\"^(${dedicated})$\"}"
           for ((i=0; i<SHARD_COUNT; i++)); do
-            matrix="${matrix},{\"shard\":$((i + 4)),\"run\":\"^(${shards[$i]})$\",\"onchain_allocation\":\"\",\"disable_activator\":\"\"}"
+            matrix="${matrix},{\"shard\":$((i + 2)),\"run\":\"^(${shards[$i]})$\"}"
           done
           matrix="${matrix}]"
 
@@ -172,7 +164,6 @@ jobs:
           pull_with_retry ${{ env.DZ_IMAGE_REPO }}/base:${{ env.DZ_IMAGE_TAG }}
           pull_with_retry ${{ env.DZ_IMAGE_REPO }}/ledger:${{ env.DZ_IMAGE_TAG }}
           pull_with_retry ${{ env.DZ_IMAGE_REPO }}/controller:${{ env.DZ_IMAGE_TAG }}
-          pull_with_retry ${{ env.DZ_IMAGE_REPO }}/activator:${{ env.DZ_IMAGE_TAG }}
           pull_with_retry ${{ env.DZ_IMAGE_REPO }}/manager:${{ env.DZ_IMAGE_TAG }}
           pull_with_retry ${{ env.DZ_IMAGE_REPO }}/funder:${{ env.DZ_IMAGE_TAG }}
           pull_with_retry ${{ env.DZ_IMAGE_REPO }}/device:${{ env.DZ_IMAGE_TAG }}
@@ -186,6 +177,4 @@ jobs:
       - name: test
         env:
           DZ_E2E_NO_BUILD: "1"
-          DZ_E2E_ONCHAIN_ALLOCATION: ${{ matrix.onchain_allocation }}
-          DZ_E2E_DISABLE_ACTIVATOR: ${{ matrix.disable_activator }}
         run: go test -tags=e2e -timeout=20m -parallel=12 -run '${{ matrix.run }}' -v
