+++ e2e/.env.local
@@ -5,7 +5,6 @@ DZ_BASE_IMAGE=${DZ_IMAGE_REPO}/base:${DZ_IMAGE_TAG}
 
 DZ_LEDGER_IMAGE=${DZ_IMAGE_REPO}/ledger:${DZ_IMAGE_TAG}
 DZ_CONTROLLER_IMAGE=${DZ_IMAGE_REPO}/controller:${DZ_IMAGE_TAG}
-DZ_ACTIVATOR_IMAGE=${DZ_IMAGE_REPO}/activator:${DZ_IMAGE_TAG}
 DZ_MANAGER_IMAGE=${DZ_IMAGE_REPO}/manager:${DZ_IMAGE_TAG}
 DZ_FUNDER_IMAGE=${DZ_IMAGE_REPO}/funder:${DZ_IMAGE_TAG}
 DZ_DEVICE_IMAGE=${DZ_IMAGE_REPO}/device:${DZ_IMAGE_TAG}
