+++ e2e/internal/devnet/ledger.go
@@ -97,7 +97,7 @@ func (l *Ledger) ExternalRPCURL() string {
 	return fmt.Sprintf("http://%s:%d", l.dn.ExternalHost, l.ExternalRPCPort)
 }
 
-// dockerContainerName returns the name of the deterministic activator container based on the
+// dockerContainerName returns the name of the deterministic ledger container based on the
 // deployID and component name.
 func (l *Ledger) dockerContainerName() string {
 	return l.dn.Spec.DeployID + "-" + l.dockerContainerHostname()
